fixes from PR comments
