+++ deepchecks/tabular/checks/model_evaluation/weak_segments_performance.py
@@ -15,7 +15,7 @@ from typing import TYPE_CHECKING, Callable, Dict, List, Union
 import numpy as np
 import pandas as pd
 
-from deepchecks.core import CheckResult, ConditionCategory, ConditionResult
+from deepchecks.core import CheckResult
 from deepchecks.core.check_result import DisplayMap
 from deepchecks.core.errors import DeepchecksNotSupportedError, DeepchecksProcessError, DeepchecksValueError
 from deepchecks.tabular import Context, SingleDatasetCheck
@@ -24,7 +24,6 @@ from deepchecks.tabular.utils.task_type import TaskType
 from deepchecks.utils.docref import doclink
 from deepchecks.utils.performance.weak_segment_abstract import WeakSegmentAbstract
 from deepchecks.utils.single_sample_metrics import calculate_per_sample_loss
-from deepchecks.utils.strings import format_number, format_percent
 from deepchecks.utils.typing import Hashable
 
 if TYPE_CHECKING:
@@ -172,26 +171,3 @@ class WeakSegmentsPerformance(SingleDatasetCheck, WeakSegmentAbstract):
                         f'{reference}Scorer name: {k}'
                     )
         return super().config(include_version)
-
-    def add_condition_segments_relative_performance_greater_than(self, max_ratio_change: float = 0.20):
-        """Add condition - check that the score of the weakest segment is greater than supplied relative threshold.
-
-        Parameters
-        ----------
-        max_ratio_change : float , default: 0.20
-            maximal ratio of change allowed between the average score and the score of the weakest segment.
-        """
-
-        def condition(result: Dict) -> ConditionResult:
-            weakest_segment_score = result['weak_segments_list'].iloc[0, 0]
-            msg = f'Found a segment with {result["scorer_name"]} score of {format_number(weakest_segment_score, 3)} ' \
-                  f'in comparison to an average score of {format_number(result["avg_score"], 3)} in sampled data.'
-            if result['avg_score'] > 0 and weakest_segment_score > (1 - max_ratio_change) * result['avg_score']:
-                return ConditionResult(ConditionCategory.PASS, msg)
-            elif result['avg_score'] < 0 and weakest_segment_score > (1 + max_ratio_change) * result['avg_score']:
-                return ConditionResult(ConditionCategory.PASS, msg)
-            else:
-                return ConditionResult(ConditionCategory.WARN, msg)
-
-        return self.add_condition(f'The relative performance of weakest segment is greater than '
-                                  f'{format_percent(1 - max_ratio_change)} of average model performance.', condition)
